pyproject.toml
[project]

name = "cryptostream"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Braga",email = "lucassudbrackb@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"

dependencies = [
    "dbt-bigquery (>=1.10.3,<2.0.0)",
    "dbt-core (>=1.10.15,<2.0.0)",
    "apache-airflow (>=3.1.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
